requirements.txt
google-genai

numpy

beautifulsoup4

requests

lxml

urllib3

python-dateutil

pyyaml
